+++ src/main/java/CorruptFileException.java
@@ -1,0 +1,5 @@
+public class CorruptFileException extends Exception {
+    public CorruptFileException() {
+        super("Save file is corrupt.");
+    }
+}
