+++ src/main/java/Pookie.java
@@ -2,14 +2,13 @@ import java.io.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Pookie {
     private static final String FILE_PATH = "./data/pookie.txt";
     private static final File FILE = new File(FILE_PATH);
-    private static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
-    private static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mma");
+    public static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
+    public static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mma");
 
     private Storage storage;
     private TaskList tasks;
@@ -19,214 +18,6 @@ public class Pookie {
         this.storage = storage;
         this.tasks = new TaskList(storage.loadTasks());
         this.testMode = testMode;
-    }
-
-    static class CorruptFileException extends Exception {
-        public CorruptFileException() {
-            super("Save file is corrupt.");
-        }
-    }
-
-    static abstract class Task {
-        private boolean isDone;
-        private String description;
-
-        public Task(Boolean isDone, String description) {
-            this.isDone = isDone;
-            this.description = description;
-        }
-
-        public boolean getIsDone() {
-            return isDone;
-        }
-
-        public String getDescription() {
-            return description;
-        }
-
-        public String getStatusIcon() {
-            return (isDone ? "X" : " "); // mark done task with X
-        }
-
-        public void markAsDone() {
-            this.isDone = true;
-        }
-
-        public void markAsNotDone() {
-            this.isDone = false;
-        }
-
-        @Override
-        public String toString() {
-            return "[ ] " + description;
-        }
-
-        public abstract String toFileString();
-    }
-
-    static class Todo extends Task {
-        public Todo(Boolean isDone, String description) {
-            super(isDone, description);
-        }
-
-        @Override
-        public String toString() {
-            return "[T][" + getStatusIcon() + "] " + getDescription();
-        }
-
-        @Override
-        public String toFileString() {
-            return "T | " + (getIsDone() ? 1 : 0) + " | " + getDescription();
-        }
-    }
-
-    static class Deadline extends Task {
-        private LocalDateTime by;
-
-        public Deadline(Boolean isDone, String description, LocalDateTime by) {
-            super(isDone, description);
-            this.by = by;
-        }
-
-        @Override
-        public String toString() {
-            return "[D][" + getStatusIcon() + "] " + getDescription() + " (by: " + by.format(OUTPUT_FORMATTER) + ")";
-        }
-
-        @Override
-        public String toFileString() {
-            return "D | " + (getIsDone() ? 1 : 0) + " | " + getDescription() + " | " + by.format(OUTPUT_FORMATTER);
-        }
-    }
-
-    static class Event extends Task {
-        private LocalDateTime from;
-        private LocalDateTime to;
-
-        public Event(Boolean isDone, String description, LocalDateTime from, LocalDateTime to) {
-            super(isDone, description);
-            this.from = from;
-            this.to = to;
-        }
-
-        @Override
-        public String toString() {
-            return "[E][" + getStatusIcon() + "] " + getDescription() + " (from: " + from.format(OUTPUT_FORMATTER) + " to: " + to.format(OUTPUT_FORMATTER) + ")";
-        }
-
-        @Override
-        public String toFileString() {
-            return "E | " + (getIsDone() ? 1 : 0) + " | " + getDescription() + " | " + from.format(OUTPUT_FORMATTER) + " | " + to.format(OUTPUT_FORMATTER);
-        }
-    }
-
-    static class Storage {
-        private File file;
-
-        public Storage(File file) {
-            this.file = file;
-        }
-
-        public ArrayList<Task> loadTasks() throws CorruptFileException, IOException {
-            try (BufferedReader br = new BufferedReader(new FileReader(file))) {
-                ArrayList<Task> tasks = new ArrayList<>();
-                String line;
-                while ((line = br.readLine()) != null) {
-                    String[] parts = line.split(" \\| ");
-                    if (parts.length < 3) throw new CorruptFileException();
-
-                    if (!parts[1].equals("1") && !parts[1].equals("0")) {
-                        throw new CorruptFileException();
-                    }
-                    boolean isDone = parts[1].equals("1");
-                    String description = parts[2];
-
-                    Task task = null;
-                    switch (parts[0]) {
-                        case "T":
-                            if (parts.length != 3) throw new CorruptFileException();
-                            task = new Todo(isDone, description);
-                            break;
-                        case "D":
-                            if (parts.length != 4) throw new CorruptFileException();
-                            String byStr = parts[3];
-                            LocalDateTime by = null;
-                            try {
-                                by = LocalDateTime.parse(byStr, OUTPUT_FORMATTER);
-                            } catch (DateTimeParseException e) {
-                                throw new CorruptFileException();
-                            }
-                            task = new Deadline(isDone, description, by);
-                            break;
-                        case "E":
-                            if (parts.length != 5) throw new CorruptFileException();
-                            String fromStr = parts[3];
-                            String toStr = parts[4];
-                            LocalDateTime from = null;
-                            LocalDateTime to = null;
-                            try {
-                                from = LocalDateTime.parse(fromStr, OUTPUT_FORMATTER);
-                                to = LocalDateTime.parse(toStr, OUTPUT_FORMATTER);
-                            } catch (DateTimeParseException e) {
-                                throw new CorruptFileException();
-                            }
-                            task = new Event(isDone, description, from, to);
-                            break;
-                        default:
-                            throw new CorruptFileException();
-                    }
-                    tasks.add(task);
-                }
-                return tasks;
-            } catch (FileNotFoundException e) {
-                return new ArrayList<>();
-            }
-        }
-
-        public void saveTasks(ArrayList<Task> tasks) throws IOException {
-            file.getParentFile().mkdirs();
-            PrintWriter pw = new PrintWriter(new FileWriter(file));
-            for (Task task : tasks) {
-                pw.println(task.toFileString());
-            }
-            pw.close();
-        }
-    }
-
-    static class TaskList {
-        private ArrayList<Task> tasks;
-
-        public TaskList() {
-            this.tasks = new ArrayList<>();
-        }
-
-        public TaskList(ArrayList<Task> tasks) {
-            this.tasks = tasks;
-        }
-
-        public Task get(int index) {
-            return tasks.get(index);
-        }
-
-        public void add(Task task) {
-            tasks.add(task);
-        }
-
-        public Task remove(int index) {
-            return tasks.remove(index);
-        }
-
-        public int size() {
-            return tasks.size();
-        }
-
-        public boolean isEmpty() {
-            return tasks.isEmpty();
-        }
-
-        public ArrayList<Task> getList() {
-            return new ArrayList<>(tasks);
-        }
     }
 
     public static void main(String[] args) throws CorruptFileException, IOException {
